+++ README.md
@@ -1,72 +1,45 @@
 # kiss-icp-converter
-STEP 1 bag->kiss
 
-Nagrywanie danych z tematów 
-ros2 bag record -o bag_kiss /kiss/frame /kiss/odometry
 
-Uruchamia launch file, przekazując ścieżkę do pliku .bag i nazwę tematu.
-ros2 launch kiss_icp odometry.launch.py bagfile:=(dir_bag) topic:=(topic_name)
+## Intended use 
 
-STEP 2 bag -> session
+This small toolset allows to integrate SLAM solution provided by [kiss-icp](https://github.com/PRBonn/kiss-icp) with [HDMapping](https://github.com/MapsHD/HDMapping).
+This repository contains ROS 2 workspace that :
+  - submodule to tested revision of Kiss ICP
+  - a converter that listens to topics advertised from odometry node and save data in format compatible with HDMapping.
 
-Tworzenie Folder
-mkdir ros2_ws
 
-Przejscie do ros2_ws
-cd ros2_ws
+## Building
 
-klon z repo(z kropką na końcu)
-git clone https://github.com/marcinmatecki/kiss-icp-converter.git .
-
-Inicjuje submoduły w repozytorium.
-git submodule init
-
-Pobiera zawartość submodułów, w tym zagnieżdżone submoduły.
-git submodule update --recursive
-
-Kompilacja projektu
+Clone the repo
+```shell
+mkdir -p /test_ws/src
+cd /test_ws/src
+git clone https://github.com/marcinmatecki/kiss-icp-converter.git --recursive
+cd ..
 colcon build
+```
 
-Uruchomienie
+## Usage - data SLAM:
 
-ros2 run cpp_pubsub listener input_dir output_dir
+Prepare recorded bag with estimated odometry:
 
-STEP 1: bag -> kiss
+In first terminal record bag:
+```shell
+ros2 bag record /kiss/local_map /kiss/odometry
+```
 
-Recording data from topics
+and start odometry:
+```shell 
+cd /test_ws/
+source ./install/setup.sh # adjust to used shell
+ros2 launch kiss_icp odometry.launch.py bagfile:=<path_to_rosbag> topic:=<topic_name>
+```
 
-ros2 bag record -o bag_kiss /kiss/frame /kiss/odometry
+## Usage - conversion:
 
-Running the launch file, passing the path to the .bag file and topic name
-
-ros2 launch kiss_icp odometry.launch.py bagfile:=(dir_bag) topic:=(topic_name)
-
-STEP 2: bag -> session
-
-Creating a folder
-
-mkdir ros2_ws
-
-Changing to ros2_ws directory
-
-cd ros2_ws
-
-Cloning the repository (with a dot at the end)
-
-git clone https://github.com/marcinmatecki/kiss-icp-converter.git .
-
-Initializing submodules in the repository
-
-git submodule init
-
-Fetching the content of the submodules, including nested submodules
-
-git submodule update --recursive
-
-Building the project
-
-colcon build
-
-Running
-
-ros2 run cpp_pubsub listener input_dir output_dir
+```shell
+cd /test_ws/
+source ./install/setup.sh # adjust to used shell
+ros2 run kiss-icp-to-hdmapping listener <recorded_bag> <output_dir>
+```
