+++ src/kiss-icp-to-hdmapping/src/kiss_odometry_subscriber.cpp
@@ -5,7 +5,6 @@
 #include "pcl/point_types.h"
 #include "nav_msgs/msg/odometry.hpp"
 #include "pcl/conversions.h"
-#include <laszip/laszip_api.h>
 #include <fstream> 
 #include <iostream>
 #include <string>
@@ -14,16 +13,8 @@
 #include <nlohmann/json.hpp>
 #include "rosbag2_cpp/reader.hpp"
 #include "rclcpp/serialization.hpp"
+#include "laz_writer.hpp"
 
-struct Point3Di
-{
-	Eigen::Vector3d point;
-	double timestamp;
-    float intensity;
-    int index_pose;
-    uint8_t lidarid;
-	int index_point;
-};
 
 struct TrajectoryPose
 {
@@ -44,165 +35,6 @@ std::vector<Point3Di> points_global;
 std::vector<TrajectoryPose> trajectory;
 std::vector<std::vector<TrajectoryPose>> chunks_trajectory;
 
-bool saveLaz(const std::string &filename, const std::vector<Point3Di> &points_global)
-{
-
-    constexpr float scale = 0.0001f; // one tenth of milimeter
-    // find max
-    double max_x{std::numeric_limits<double>::lowest()};
-    double max_y{std::numeric_limits<double>::lowest()};
-    double max_z{std::numeric_limits<double>::lowest()};
-    double min_x = 1000000000000.0;
-    double min_y = 1000000000000.0;
-    double min_z = 1000000000000.0;
-
-    for (auto &p : points_global)
-    {
-        if (p.point.x() < min_x)
-        {
-            min_x = p.point.x();
-        }
-        if (p.point.x() > max_x)
-        {
-            max_x = p.point.x();
-        }
-
-        if (p.point.y() < min_y)
-        {
-            min_y = p.point.y();
-        }
-        if (p.point.y() > max_y)
-        {
-            max_y = p.point.y();
-        }
-
-        if (p.point.z() < min_z)
-        {
-            min_z = p.point.z();
-        }
-        if (p.point.z() > max_z)
-        {
-            max_z = p.point.z();
-        }
-    }
-
-    std::cout << "processing: " << filename << "points " << points_global.size() << std::endl;
-
-    laszip_POINTER laszip_writer;
-    if (laszip_create(&laszip_writer))
-    {
-        fprintf(stderr, "DLL ERROR: creating laszip writer\n");
-        return false;
-    }
-
-    // get a pointer to the header of the writer so we can populate it
-
-    laszip_header *header;
-
-    if (laszip_get_header_pointer(laszip_writer, &header))
-    {
-        fprintf(stderr, "DLL ERROR: getting header pointer from laszip writer\n");
-        return false;
-    }
-
-    // populate the header
-
-    header->file_source_ID = 4711;
-    header->global_encoding = (1 << 0); // see LAS specification for details
-    header->version_major = 1;
-    header->version_minor = 2;
-    //    header->file_creation_day = 120;
-    //    header->file_creation_year = 2013;
-    header->point_data_format = 1;
-    header->point_data_record_length = 0;
-    header->number_of_point_records = points_global.size();
-    header->number_of_points_by_return[0] = points_global.size();
-    header->number_of_points_by_return[1] = 0;
-    header->point_data_record_length = 28;
-    header->x_scale_factor = scale;
-    header->y_scale_factor = scale;
-    header->z_scale_factor = scale;
-
-    header->max_x = max_x;
-    header->min_x = min_x;
-    header->max_y = max_y;
-    header->min_y = min_y;
-    header->max_z = max_z;
-    header->min_z = min_z;
-
-    // optional: use the bounding box and the scale factor to create a "good" offset
-    // open the writer
-    laszip_BOOL compress = (strstr(filename.c_str(), ".laz") != 0);
-
-    if (laszip_open_writer(laszip_writer, filename.c_str(), compress))
-    {
-        fprintf(stderr, "DLL ERROR: opening laszip writer for '%s'\n", filename.c_str());
-        return false;
-    }
-
-    fprintf(stderr, "writing file '%s' %scompressed\n", filename.c_str(), (compress ? "" : "un"));
-
-    // get a pointer to the point of the writer that we will populate and write
-
-    laszip_point *point;
-    if (laszip_get_point_pointer(laszip_writer, &point))
-    {
-        fprintf(stderr, "DLL ERROR: getting point pointer from laszip writer\n");
-        return false;
-    }
-
-    laszip_I64 p_count = 0;
-    laszip_F64 coordinates[3];
-
-    for (int i = 0; i < points_global.size(); i++)
-    {
-        const auto &p = points_global[i];
-        point->intensity = p.intensity;
-        point->gps_time = p.timestamp * 1e9;
-        p_count++;
-        coordinates[0] = p.point.x();
-        coordinates[1] = p.point.y();
-        coordinates[2] = p.point.z();
-        if (laszip_set_coordinates(laszip_writer, coordinates))
-        {
-            fprintf(stderr, "DLL ERROR: setting coordinates for point %I64d\n", p_count);
-            return false;
-        }
-
-        if (laszip_write_point(laszip_writer))
-        {
-            fprintf(stderr, "DLL ERROR: writing point %I64d\n", p_count);
-            return false;
-        }
-    }
-
-    if (laszip_get_point_count(laszip_writer, &p_count))
-    {
-        fprintf(stderr, "DLL ERROR: getting point count\n");
-        return false;
-    }
-
-    fprintf(stderr, "successfully written %I64d points\n", p_count);
-
-    // close the writer
-
-    if (laszip_close_writer(laszip_writer))
-    {
-        fprintf(stderr, "DLL ERROR: closing laszip writer\n");
-        return false;
-    }
-
-    // destroy the writer
-
-    if (laszip_destroy(laszip_writer))
-    {
-        fprintf(stderr, "DLL ERROR: destroying laszip writer\n");
-        return false;
-    }
-
-    std::cout << "exportLaz DONE" << std::endl;
-    return true;
-}
 
 bool save_poses(const std::string file_name, std::vector<Eigen::Affine3d> m_poses, std::vector<std::string> filenames)
 {
